# Models/SendToGREET1/CanolaHEFA.cs
namespace GREETApi.Models.SendToGREET1
{
    public class CanolaHEFA : CommonHEFAFields
    {
        public decimal AgTractor_BDShare_Canola { get; set; }
        public decimal AgTractor_ConvDieselShare_Canola { get; set; }
        public decimal AgTractor_RDShare_Canola { get; set; }
        public decimal Blend_Type_Canola { get; set; }
        public decimal Canola_HEFA_Electricity { get; set; }
        public decimal Canola_HEFA_Hydrogen { get; set; }
        public decimal Canola_HEFA_Natural_gas { get; set; }
        public decimal Canola_Oil_Extraction_Allocation { get; set; }
        public decimal Green_Ammonia_Share { get; set; }
        public decimal HEFA_Allocation { get; set; }
        public decimal Motor_Type_Canola { get; set; }
    }
}